refactor: modularize agent_service.py into specialized modules
Split monolithic agent_service.py (457 lines) into 5 focused modules:

**New Module Structure:**

1. **agent_config.py** (97 lines)
   - Model routing and configuration (Ollama, Claude, Cursor)
   - Command template generation
   - Agent-to-task-type mapping
   - Executor selection (Claude/Cursor)

2. **task_store.py** (137 lines)
   - In-memory task storage (CRUD operations)
   - Task creation with model/executor routing
   - Task lifecycle management (pending → running → completed)
   - ID generation and timestamp handling

3. **task_analytics.py** (80 lines)
   - Task counting and summaries
   - Attention task filtering (needs_decision, failed)
   - Usage statistics per model
   - Route information queries

4. **pipeline_monitor.py** (186 lines)
   - Pipeline status reporting
   - Attention flags (stuck, repeated_failures, low_success_rate)
   - Running/pending/completed task tracking
   - Phase coverage by task_type

5. **agent_service.py** (57 lines) - **Slim Facade**
   - Imports and re-exports from specialized modules
   - Maintains backward compatibility
   - Clear __all__ exports

**Benefits:**

✓ **Reduced cognitive load**: 80-186 lines per module vs 457 in one file
✓ **Improved testability**: Each module independently testable
✓ **Better organization**: Clear separation of concerns
✓ **Easier maintenance**: Changes localized to specific concerns
✓ **Parallel development**: Multiple devs can work on different modules
✓ **Future-proof**: Easy to replace in-memory store with DB layer

**Backward Compatibility:**

All existing imports work unchanged:
```python
from app.services import agent_service
# All functions available via agent_service.create_task(), etc.
```

**Testing:**

- Verified all router imports remain functional
- No API changes, purely internal refactoring
- Existing tests should pass without modification
